update SPT tests
